+++ backend/src/main/java/com/festago/auth/config/LoginConfig.java
@@ -1,11 +1,11 @@
 package com.festago.auth.config;
 
 import com.festago.auth.domain.AuthExtractor;
-import com.festago.auth.domain.AuthInterceptor;
-import com.festago.auth.domain.AuthenticateContext;
 import com.festago.auth.domain.Role;
 import com.festago.auth.infrastructure.CookieTokenExtractor;
 import com.festago.auth.infrastructure.HeaderTokenExtractor;
+import com.festago.auth.presentation.AuthInterceptor;
+import com.festago.auth.presentation.AuthenticateContext;
 import com.festago.auth.presentation.RoleArgumentResolver;
 import java.util.List;
 import org.springframework.context.annotation.Bean;
